+++ src/pages/mypage/bookmark/index.page.tsx
@@ -9,7 +9,7 @@ import HeaderLayout from "@/components/Layout/HeaderLayout";
 import { BOOKMARK_KEYS } from "@/constants/queryKey";
 
 import * as S from "./styled";
-import SkeletonGrid from "../components/Bookmark/Skeleton/SkeletonGrid/index";
+import SkeletonGrid from "../../mypage/components/Bookmark/Skeleton/SkeletonGrid";
 
 export const getServerSideProps = async () => {
   const queryClient = new QueryClient();
